nasmlib: Plcate undefined behaviour in signed_bits
http://bugzilla.nasm.us/show_bug.cgi?id=3392320

Reported-by: Dingbao Xie <xiedingbao@gmail.com>

+++ nasmlib.h
@@ -456,7 +456,7 @@ static inline int64_t signed_bits(int64_t value, int bits)
     if (bits < 64) {
         value &= ((int64_t)1 << bits) - 1;
         if (value & (int64_t)1 << (bits - 1))
-            value |= (int64_t)-1 << bits;
+            value |= (int64_t)((uint64_t)-1 << bits);
     }
     return value;
 }
